Bug fixes and adaptions
I followed Kasias list of what to change. In detail:
• Load image from tiff – remove
• Time between frame – default is 1 sec maybe better change to sth like
1 h
• Remove channel choice green red etc since it is not used
• Edge detection should be unticked on default
• Edge detection and correlation parameters should be in pixels
• Interpolation should be in um hours etc
• Remove the white box on the top right
• What about the small image on the bottom right?
• Add the preview of color map to the interface
• There is still sometimes problem with the scale bar. After couple of
images with right scale bar it changes it to sth wired – not the right
scale (see an example). In the example below the first scale bar is ok,
remains ok for 6 images and then changes to sth else
• Folders:
o Remove plus actin folder
o Rename the folders (maybe, if complictated i think it will be fine
also to leave as it is): we have gray scale so maybe instead of
  filter_overlay_im to interpolation_image
 filter_overlay_data to interpolation_data
 “image” to correlation_image?
o If possible can we separate the images within the filter_overlay_im
into two folders: one for the images with scale bar and the other with
the images with just colours (this i am not sure if we need in fact as
we have gray scale images now). Originally i think the idea was to use
the colour images with no arrow to measure the contraction average but
we don’t really do it. So I would either remove. If necessary we can
have two folders with interpolation image -colour and arrow: one with
scale bar and the other without. Not sure if this is necessary but
maybe.
• Also name of the files maybe it is better to name everything the same
way like say my png is called E32S1 then it would be:
E32S1_grayscale_0000, 0001 etc and also E32S1_interpolation_0000. Again
if to much work don’t worry.
• The scale bar image is kind of squashed on the images. I think it
keeps the original size of the image and then stick the bar next to it,
which makes the original image decrease in width but the length is the
same. Maybe its better to expand it. I don’t know if possible. See below
one is gray scale not squeezed and the one with bar is squeezed.
I will add the scalebar as extra file, then it should be ok.